Fixed: wflow-wfengine - Deadline and due date are calculated wrongly when value is larger than 25 days.  #2692 @trunk
git-svn-id: http://dev.joget.org/svn/jw-community/trunk@1839 1bff935d-2fe2-40b7-b00f-c2aca69a7612

+++ wflow-wfengine/src/main/java/org/joget/workflow/model/WorkflowDeadline.java
@@ -6,15 +6,15 @@ import java.util.regex.Pattern;
 
 public class WorkflowDeadline {
 
-    private int deadlineLimit;
+    private long deadlineLimit;
     private String deadlineExpression;
     private Map context;
 
-    public int getDeadlineLimit() {
+    public long getDeadlineLimit() {
         return deadlineLimit;
     }
 
-    public void setDeadlineLimit(int deadlineLimit) {
+    public void setDeadlineLimit(long deadlineLimit) {
         this.deadlineLimit = deadlineLimit;
         setExpressionFromLimit();
     }
